requirements.txt
matplotlib
tkcalender
pillow
urllib3==1.1
